feat(ui): add help page and interaction sound
chore(cleanup): remove unused Vue files

refactor(router): update routejs
