--- src/sidebar/router/pages/VoiceInput.vue
@@ -1,4 +1,0 @@
-<!-- Redirects to here from Index after user clicks on "Voice Input"    -->
-<!-- TODO:  Add a message to request user to speak certain message      -->
-<!--        When user is done speaking, ask to click button             -->
-<!--        This button redirects to Evaluation page with new results   -->
